Remove pip call from main thread
The problem we have with calling pip from the main thread is that
in some cases, pip may fail to complete and segfault while installing
some dependencies. This in return causes a swarm build of docker
to get stuck into a boot loop and docker doesn't detect the failure
right away and still think process converged.

A longer wait period could help detect failures but it's difficult to
tell what caused the issue. Regardless of that, the server may still
be able to boot as segfault may be caused by installing twice a
dependency. For example, the cryptography module may segfault pip
if already installed. But it doesn't prevent odoo from working
correctly.

Thus, we allow odoo from booting even if the pip phase failed. A decent
monitoring solution should raise the issue and if it would prevent
odoo from working correctly, a decent monitoring solution would still
raise the issue later as it would be correctly shown in the logs that
someting failed.