Update entrypoint. Wait for postgres
2 new environment variables are added

PGRETRY = the amount of time you want to retry (default: 5)
PGRETRYTIME = the amount to wait between retries

Updated the entrypoint to wait for postgres. The connection uses environmnent
variables that are defined + pgpass if available.

Connection string is empty to use default values, this way, you can
run odoo without having to pass a password directly to odoo. It will
be used by postgres and the password can remain encrypted as a secret
file.

So credentials are more or less protected even if copying the password
from .pgpass should be enough for anyone to connect from somewhere else
if postgres is badly configured.