Added tiny version of odoo
Odoo is installed directly in /var/lib/odoo which is defined as a
docker volume. The volumes will automatically copy the odoo installation.

If you mount /var/lib/odoo as a volume, you'll be able to get odoo
installed directly in your home and then it will be possible to update
your odoo installation by upgrading odoo using pip.

The advantage is that all the non required dependencies are removed
you can virtually update the odoo code to any version and the ready
version of odoo installed is just a way to bootstrap a more complex
odoo setup in the future.