Add todo list to readme
The build process can be largely improved, as I found out recently,
it's possible that some build will actually fail unit tests of odoo

The previous release of Odoo 12.0 would actually fail a few tests,
in the future, it would make more sense to pass the job of building
image to a CI system that will be able to build the image
then test the image then deploy the image if unit tests actually pass.

Each of the new images should be released with their own specific tag
while the 12.0, 13.0 etc are going to be manually updated.

In the end there will be 3 categories of tags:

- X.0 : Manually updated from time to time (may be monthly)
- X.0-YYYYMMDD: Automatically updated whenever CI test pass
- X.0-nightly: Automatically updated even if CI tests fail