Improved build script to make it easier to prepare new images
The build script can be used to directly push to a private registry
instead of index.docker.io. It's possible to build images locally
and store the computed dockerfile assets etc directly from the script.

This is a plan To move away from storing the Docker file directly in
the project. It appears that some people fork the project and customize
the built images instead of modifying the assets. The assets and templates
should be modified in order to be able to build/improve even older version
of odoo.

It's technically possible to build any release anytime based on the
versions.toml file.

In the long run, it will be possible to remove the docker hub automatic
image rebuild to trigger a build on a different server only when necessary
instead of daily.

Or even build when a new release is pushed to nightly.odoo.com